refactor: 컨트롤러 1개로 통합 및 ChangeLogDiff -> ChangeLogDiffResponse로 이름 변경
